Reverse VERIFY_PEER exploit Commit
I don't think it made any sense to merge this commit
https://github.com/AuthorizeNet/sdk-php/commit/45b647b3778d14cd761003342c25e438aa2bed06

You have a setting, that if you set it to false, then it just returns nothing. I am using this in a test environment and don't want to setup SSL for the server, so I should be able to turn off VERIFY_PEER at my own risk. Do not make the code hand hold people and prevent them from having a built in option for the curl extension.

Thanks.