Add proper markdown to format code blocks
I added the required markdown to format the sections of code as proper code blocks. I'm currently trying to implement this feature in my site, and the documentation is painful to read.

I'm only using this doc page at the moment so I haven't formatted the others, but I'd be happy to do that when I have a minute if you'd like.